+++ ckanext/harvest/logic/action/update.py
@@ -9,18 +9,17 @@ import datetime
 
 from ckantoolkit import config
 from sqlalchemy import and_, or_
+from six.moves.urllib.parse import urljoin
 
 from ckan.lib.search.index import PackageSearchIndex
-from ckan.plugins import PluginImplementations
+from ckan.plugins import toolkit, PluginImplementations, IActions
 from ckan.logic import get_action
 from ckanext.harvest.interfaces import IHarvester
 from ckan.lib.search.common import SearchIndexError, make_connection
-
+from ckan.lib.base import render_jinja2
 
 from ckan.model import Package
 from ckan import logic
-from ckan.plugins import toolkit
-
 
 from ckan.logic import NotFound, check_access
 
@@ -234,6 +233,80 @@ def harvest_source_clear(context, data_dict):
     get_action('harvest_source_reindex')(context, {'id': harvest_source_id})
 
     return {'id': harvest_source_id}
+
+
+def harvest_abort_failed_jobs(context, data_dict):
+    session = context['session']
+
+    try:
+        life_span = int(data_dict.get('life_span'))
+    except ValueError:
+        life_span = 7
+
+    include_sid = []
+    exclude_sid = []
+    include = data_dict.get('include')
+
+    if include:
+        include_sid = set(_id for _id in include.split(','))
+
+    # TODO: if included sources provided do we want to use exclude?
+    if data_dict.get('exclude') and not include:
+        exclude_sid = set((_id for _id in data_dict.get('exclude').split(',')))
+
+    # lifespan is based on source update frequency
+    update_map = {
+        'DAILY': 1,
+        'WEEKLY': 7,
+        'BIWEEKLY': 14,
+        'MONTHLY': 30,
+        'MANUAL': life_span,
+        'ALWAYS': life_span
+    }
+
+    current_time = datetime.datetime.utcnow()
+
+    # get all running jobs
+    jobs_list = session.query(HarvestJob.id,
+                              HarvestJob.source_id,
+                              HarvestJob.created) \
+        .filter(HarvestJob.status == 'Running') \
+        .all()
+
+    # filter out not included source job's
+    if include_sid:
+        jobs_list = [
+            job for job in jobs_list
+            if job.source_id in include_sid
+        ]
+
+    if not jobs_list:
+        return 'There is no jobs to abort'
+
+    aborted_counter = 0
+    for job in jobs_list:
+        harvest_source = session.query(HarvestSource.frequency) \
+                                .filter(HarvestSource.id == job.source_id) \
+                                .first()
+
+        life_span = update_map.get(harvest_source.frequency)
+        if not life_span:
+            raise Exception('Frequency {freq} not recognised'.format(
+                freq=harvest_source.frequency))
+
+        expire_date = current_time - datetime.timedelta(days=life_span)
+
+        if not include_sid and job.source_id in exclude_sid:
+            log.info('Excluding source: {}'.format(job.source_id))
+            continue
+        # if job is running too long, abort it
+        if job.created < expire_date:
+            log.info(get_action('harvest_job_abort')(context, {'id': job.id}))
+            aborted_counter += 1
+        else:
+            log.info('{} running less then {} days. Skipping...'.format(job.id, life_span))
+    else:
+        return 'Done. Aborted jobs: {}'.format(aborted_counter)
 
 
 def harvest_sources_job_history_clear(context, data_dict):
@@ -537,14 +610,16 @@ def harvest_jobs_run(context, data_dict):
         context, {'source_id': source_id, 'status': u'Running'})
     if len(jobs):
         for job in jobs:
+            job_obj = HarvestJob.get(job['id'])
             if timeout:
-                created = datetime.datetime.strptime(job['created'], '%Y-%m-%d %H:%M:%S.%f')
+                last_time = job_obj.get_last_action_time()
                 now = datetime.datetime.now()
-                if now - created > datetime.timedelta(minutes=int(timeout)):
-                    msg = 'Job timeout: %s is taking longer than %s minutes' % (job['id'], timeout)
-                    log.error(msg)
-
-                    job_obj = HarvestJob.get(job['id'])
+                if now - last_time > datetime.timedelta(minutes=int(timeout)):
+                    msg = 'Job {} timeout ({} minutes)\n'.format(job_obj.id, timeout)
+                    msg += '\tJob created: {}\n'.format(job_obj.created)
+                    msg += '\tJob gather finished: {}\n'.format(job_obj.created)
+                    msg += '\tJob last action time: {}\n'.format(last_time)
+                    
                     job_obj.status = u'Finished'
                     job_obj.finished = now
                     job_obj.save()
@@ -564,7 +639,7 @@ def harvest_jobs_run(context, data_dict):
                            .count()
 
                 if num_objects_in_progress == 0:
-                    job_obj = HarvestJob.get(job['id'])
+                    
                     job_obj.status = u'Finished'
                     log.info('Marking job as finished %s %s',
                              job_obj.source.url, job_obj.id)
@@ -588,14 +663,21 @@ def harvest_jobs_run(context, data_dict):
                     get_action('harvest_source_reindex')(
                         context, {'id': job_obj.source.id})
 
-                    status = get_action('harvest_source_show_status')(context, {'id': job_obj.source.id})
+                    status = get_action('harvest_source_show_status')(
+                        context, {'id': job_obj.source.id})
 
-                    if toolkit.asbool(config.get('ckan.harvest.status_mail.errored'))\
-                            and (status['last_job']['stats']['errored']):
-                        send_error_mail(context, job_obj.source.id, status)
+                    notify_all = toolkit.asbool(config.get('ckan.harvest.status_mail.all'))
+                    notify_errors = toolkit.asbool(config.get('ckan.harvest.status_mail.errored'))
+                    last_job_errors = status['last_job']['stats'].get('errored', 0)
+                    log.debug('Notifications: All:{} On error:{} Errors:{}'.format(notify_all, notify_errors, last_job_errors))
+                    
+                    if last_job_errors > 0 and (notify_all or notify_errors):
+                        send_error_email(context, job_obj.source.id, status)
+                    elif notify_all:
+                        send_summary_email(context, job_obj.source.id, status)
                 else:
-                    log.debug('Ongoing job:%s source:%s',
-                              job['id'], job['source_id'])
+                    log.debug('%d Ongoing jobs for %s (source:%s)',
+                              num_objects_in_progress, job['id'], job['source_id'])
     log.debug('No jobs to send to the gather queue')
 
     # Resubmit old redis tasks
@@ -607,10 +689,18 @@ def harvest_jobs_run(context, data_dict):
     return []  # merely for backwards compatibility
 
 
-def send_error_mail(context, source_id, status):
-
+def get_mail_extra_vars(context, source_id, status):
     last_job = status['last_job']
+    
     source = get_action('harvest_source_show')(context, {'id': source_id})
+    report = get_action(
+        'harvest_job_report')(context, {'id': status['last_job']['id']})
+    obj_errors = []
+    job_errors = []
+    
+    for harvest_object_error_key in islice(report.get('object_errors'), 0, 20):
+        harvest_object_error = report.get(
+            'object_errors')[harvest_object_error_key]['errors']
 
     ckan_site_url = config.get('ckan.site_url')
     job_url = toolkit.url_for('harvest_job_show', source=source['id'], id=last_job['id'])
@@ -627,14 +717,15 @@ def send_error_mail(context, source_id, status):
         msg += toolkit._('Organization: {0}').format(source['organization']['name'])
         msg += '\n\n'
 
-    msg += toolkit._('Harvest Job Id: {0}').format(last_job['id']) + '\n'
-    msg += toolkit._('Created: {0}').format(last_job['created']) + '\n'
-    msg += toolkit._('Finished: {0}').format(last_job['finished']) + '\n\n'
+    for harvest_gather_error in islice(report.get('gather_errors'), 0, 20):
+        job_errors.append(harvest_gather_error['message'])
 
-    report = get_action('harvest_job_report')(context, {'id': status['last_job']['id']})
+    if source.get('organization'):
+        organization = source['organization']['name']
+    else:
+        organization = 'Not specified'
 
-    msg += toolkit._('Records in Error: {0}').format(str(last_job['stats'].get('errored', 0)))
-    msg += '\n'
+    harvest_configuration = source.get('config')
 
     msg += 'For help, please contact the NCAR Data Stewardship Coordinator (mailto:datahelp@ucar.edu).\n\n\n'
 
@@ -699,13 +790,14 @@ def send_error_mail(context, source_id, status):
                      'subject': config.get('ckan.site_title') + ' - Harvesting Job - Error Notification',
                      'body': msg}
 
-            try:
-                mailer.mail_recipient(**email)
-            except mailer.MailerException:
-                log.error('Sending Harvest-Notification-Mail failed. Message: ' + msg)
-            except Exception as e:
-                log.error(e)
-                raise
+        try:
+            mailer.mail_recipient(**email)
+        except mailer.MailerException:
+            log.error(
+                'Sending Harvest-Notification-Mail failed. Message: ' + body)
+        except Exception as e:
+            log.error(e)
+            raise
 
 
 def harvest_send_job_to_gather_queue(context, data_dict):